% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_phred_fastq.R
\name{get_phred_fastq}
\alias{get_phred_fastq}
\title{Extract phred scores from fastq file}
\usage{
get_phred_fastq(filepath, outfile)
}
\arguments{
\item{filepath}{the file path of the fastq file}

\item{outfile}{the output file path of the fastq phred scores sequence}
}
\description{
Read in the fastq file line by line. Discard all information except for phred scores.
Phred scores are output to a file
}
