% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seq_fasta.R
\name{get_seq_fasta}
\alias{get_seq_fasta}
\title{Extract sequence from fasta file}
\usage{
get_seq_fasta(filepath, outfile)
}
\arguments{
\item{filepath}{the file path of the fasta file}

\item{outfile}{the output file path of the fasta sequence}
}
\description{
Read in the fasta file line by line. Discard sequence id and output the sequence
}
