% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phred2prob.R
\name{phred2prob}
\alias{phred2prob}
\title{Convert Phred scores to Error probabilities}
\usage{
phred2prob(phred, log_scale = TRUE)
}
\arguments{
\item{phred}{phred scores}

\item{log_scale}{whether to get the probability in log10 scale or not}
}
\description{
convert the phred scores to error probability or log10(error probability)
}
\examples{
phred2prob("BE")  # -3.3 -3.6
phred2prob("BE",  log_scale = FALSE) # 0.0005011872 0.0002511886
phred2prob("!!!", log_scale = FALSE) # 1 1 1
}
